+++ platform/stm32h750/shellio.c
@@ -149,7 +149,7 @@ static void printFSError(OsFileError error)
     }
 }
 
-void shellMount(int argc, char *argv[])
+void shellMount(long argc, char *argv[])
 {
     if (argc >= 3)
     {
@@ -162,7 +162,7 @@ void shellMount(int argc, char *argv[])
     }
 }
 
-void shellUnmount(int argc, char *argv[])
+void shellUnmount(long argc, char *argv[])
 {
     if (argc >= 2)
     {
@@ -175,14 +175,14 @@ void shellUnmount(int argc, char *argv[])
     }
 }
 
-void shellPWD(int argc, char *argv[])
+void shellPWD(long argc, char *argv[])
 {
     char buffer[OS_MAX_FILE_PATH_LENGTH];
     osFGetCWD(buffer, OS_MAX_FILE_PATH_LENGTH);
     shellWriteString(&sShell, buffer);
 }
 
-void shellCD(int argc, char *argv[])
+void shellCD(long argc, char *argv[])
 {
     if (argc >= 2)
     {
@@ -294,7 +294,7 @@ static void showFileInfo(OsFileInfo *fileInfo)
     shellWriteString(&sShell, "\n");
 }
 
-void shellLS(int argc, char *argv[])
+void shellLS(long argc, char *argv[])
 {
     char *path = "";
     if (argc >= 2)
@@ -315,7 +315,7 @@ void shellLS(int argc, char *argv[])
     }
 }
 
-void shellFind(int argc, char *argv[])
+void shellFind(long argc, char *argv[])
 {
     if (argc >= 2)
     {
@@ -352,7 +352,7 @@ static void showFSInfo(const OsFS *fs)
     shellPrint(&sShell, "所有页：%lld  ", fs->totalPages);
 }
 
-void shellStatFS(int argc, char *argv[])
+void shellStatFS(long argc, char *argv[])
 {
     if (argc >= 2)
     {
@@ -377,7 +377,7 @@ static void showMountInfo(const OsMountInfo *mountInfo)
     shellPrint(&sShell, "挂载点：%s  ", mountInfo->path);
 }
 
-void shellDF(int argc, char *argv[])
+void shellDF(long argc, char *argv[])
 {
     const OsMountInfo *mountInfo = NULL;
     while (1)
@@ -402,7 +402,7 @@ void shellDF(int argc, char *argv[])
     }
 }
 
-void shellMkDir(int argc, char *argv[])
+void shellMkDir(long argc, char *argv[])
 {
     if (argc >= 2)
     {
@@ -415,7 +415,7 @@ void shellMkDir(int argc, char *argv[])
     }
 }
 
-void shellTouch(int argc, char *argv[])
+void shellTouch(long argc, char *argv[])
 {
     if (argc >= 2)
     {
@@ -433,7 +433,7 @@ void shellTouch(int argc, char *argv[])
     }
 }
 
-void shellChMod(int argc, char *argv[])
+void shellChMod(long argc, char *argv[])
 {
     if (argc >= 3)
     {
@@ -459,7 +459,7 @@ void shellChMod(int argc, char *argv[])
     }
 }
 
-void shellStat(int argc, char *argv[])
+void shellStat(long argc, char *argv[])
 {
     if (argc >= 2)
     {
@@ -477,7 +477,7 @@ void shellStat(int argc, char *argv[])
     }
 }
 
-void shellMV(int argc, char *argv[])
+void shellMV(long argc, char *argv[])
 {
     if (argc >= 3)
     {
@@ -490,7 +490,7 @@ void shellMV(int argc, char *argv[])
     }
 }
 
-void shellRM(int argc, char *argv[])
+void shellRM(long argc, char *argv[])
 {
     if (argc >= 2)
     {
@@ -503,7 +503,7 @@ void shellRM(int argc, char *argv[])
     }
 }
 
-void shellCP(int argc, char *argv[])
+void shellCP(long argc, char *argv[])
 {
     if (argc >= 3)
     {
@@ -516,7 +516,7 @@ void shellCP(int argc, char *argv[])
     }
 }
 
-void shellFree(int argc, char *argv[])
+void shellFree(long argc, char *argv[])
 {
     shellPrint(&sShell, "所有内存：%ld\n", osTotalMem());
     shellPrint(&sShell, "可用内存：%ld\n", osFreeMem());
@@ -596,7 +596,7 @@ static void *audioPlayerTask(void *arg)
     return NULL;
 }
 
-void shellPlay(int argc, char *argv[])
+void shellPlay(long argc, char *argv[])
 {
     if (argc >= 2)
     {
@@ -621,7 +621,7 @@ void shellPlay(int argc, char *argv[])
     }
 }
 
-void shellPS(int argc, char *argv[])
+void shellPS(long argc, char *argv[])
 {
     os_task_ptr ptr;
     OsTaskInfo taskInfo;
